+++ lib/memory/profiler.rb
@@ -6,4 +6,5 @@
 require_relative "profiler/version"
 require_relative "profiler/call_tree"
 require_relative "profiler/capture"
+require_relative "profiler/allocations"
 require_relative "profiler/sampler"